% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbVARsim.R
\name{gbVARsim}
\alias{gbVARsim}
\title{Simulates a generalized binary Vector Autoregressive (gbVAR) time series.}
\usage{
gbVARsim(n, const = NULL, phi, p, theta, innov = NULL, mu, n.start = 100)
}
\arguments{
\item{n}{Sample size}

\item{const}{Constant vector, Phi0}

\item{phi}{Matrix of gbVAR coefficient probabilities of multinomial
distribution with parameter 1.}

\item{p}{The order of the gbVAR process.}

\item{theta}{Matrix of innovation coefficient probabilities of multinomial
distribution with parameter 1.}

\item{innov}{d-dimensional binary innovation process}

\item{mu}{Mean vector of innovation.}

\item{n.start}{The number of initial data to be omitted. Default is 100.}
}
\value{
time series and innovation
}
\description{
Simulates a generalized binary Vector Autoregressive (gbVAR) time series.
}
\examples{
A = matrix(c(0.49, 0.35, -0.43, -0.39), nrow = 2, byrow = T)
B = diag(c(0.16, 0.18), 2)
mu_e = c(0.4, 0.8)
result = gbVARsim(n = 100, phi = A, p = 1, theta = B, mu = mu_e)
plot(result$series, cex = 0)
lines(result$series)

}
